BUGFIX fix issue #1518
the extract method must be thread safe.
only one extract method for a gene must be executed per run of analysis.
so the first call to extract parse the hmmoutput and generate a list of hits
the later call do nothing 

in search_genes
ther is only one instance of each gene but one instance can appear several times in all_genes
(the same gene can appear in different systems)
the list of genes in paraneter of searc_genes is uniqify before to start threads