add new option --previous-run
this option can be specified only on the command line (not in config file)
if this option is set.
--cfg-file, --sequence_db, --profile_suffix, --res_extract_suffix, 
--e_value_res, --ordered, --hmmer_exe can not be specified.
the values from the previous run will be used for these options.

add previous_run in config object
define a new worker in search genes
if a hmm output is corresponding to a gene is present in the previous run,
the hmm output is used instead to recompute it.
only the extract step is rerun.

close issue #1499