FIX parsing:
in previous version the genes where created as they are dicovered in system definitions
so a gene which had an other sytem reference could be created without there attribute (inter_gene_max_space, ...)
and then these attributes were never created due to the factory design.
so the systems were incomplete and inconsistent.

reslove 
- feature #1850
- feature #1630