the parsing of Hmmoutput is extremly slow due to the DB implementation based on berkleyDB
rewriting the algo to avoid the berkleyDB. a dictionnary containig one entry by sequences found in hmmoutput 
with lenght and the rank of this sequence is build at each run.

the new algorythm is ~60 times faster tahn the previous version.

the hmm is parse a first time to extract the id of sequences present in the hmmoutput
for all this sequences a dictionnary is fill with lenght and rank strored in the txsscan indexes