if a there is 2 occurences of the same gene in one definition, skip the second
in one sytem definition the same gene can apera moer than once.
in the system or as homolog/analog
the gene was created at each time and erase the previous instance in
gene_bank.
now only the fisrt occurence is created the other are skip (they have
the same properties)