:sparkles: new gene modelization
the genes are not uniq anymore. so we have to change our model.
the profil still be uniq.
so we need as many  gene as it appear in models
but only one profile (by name and by model family)
the properties max-space-inetr-genes, loner, ... is link to a gene but
the same gene can have different values in differents models
so I cut the gene in 2
- one CoreGene which is uniq and the profile is linked to this gene
- and a ModelGene which hold all other properties specific by model