:sparkles: use hmmsearch -cut-ga opt only if GA bit is in profile
the -cut-ga does not work if the GA bit score is not present in hmm
profile. futhermore hmmsearch raise an error
so by default MSF use the -cut-ga option
if the GA bit score is not present in profile MSF swith to
e-value-search threshold 9-E in hmmsearch) and raise a warning in logs
instead the --no-cut-ga is set. in this case the --e-value-search is
applied on all profiles regardless the presence of GA bit score in
profiles