:bug: fix problem in log_level parsing
log_level sould accept int or string 'info' 'debug', ... in
configuration file.

the default is info == logging.INFO = 20
but by default the type of the filed is determine from the type of
the default.
so here we need a special method to convert string in to int which is
the internal format to store the log_level.